Move ZeroconfServiceTypes to zeroconf.services.types
